backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.35.0
gunicorn==21.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.1
cryptography==41.0.7
aiosqlite==0.19.0
pydantic==1.10.13
httpx==0.25.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==1.6.3
typing-extensions==4.13.2